requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
Babel==2.5.3
bcrypt==3.1.4
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
cliff==2.11.0
cmd2==0.8.0
cryptography==2.1.4
debtcollector==1.19.0
decorator==4.2.1
deprecation==2.0
dogpile.cache==0.6.4
Fabric3==1.14.post1
idna==2.6
iso8601==0.1.12
jmespath==0.9.3
jsonpatch==1.21
jsonpointer==2.0
jsonschema==2.6.0
keystoneauth1==3.4.0
monotonic==1.4
msgpack==0.5.6
munch==2.2.0
netaddr==0.7.19
netifaces==0.10.6
openstacksdk==0.11.3
os-client-config==1.29.0
os-service-types==1.2.0
osc-lib==1.9.0
oslo.config==5.2.0
oslo.i18n==3.19.0
oslo.serialization==2.24.0
oslo.utils==3.35.0
packaging==16.8
paramiko==2.4.0
pbr==3.1.1
prettytable==0.7.2
pyasn1==0.4.2
pycparser==2.18
PyNaCl==1.2.1
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyperclip==1.6.0
python-cinderclient==3.5.0
python-glanceclient==2.9.1
python-keystoneclient==3.15.0
python-novaclient==10.1.0
python-openstackclient==3.14.0
pytz==2018.3
PyYAML==3.12
requests==2.18.4
requestsexceptions==1.4.0
rfc3986==1.1.0
simplejson==3.13.2
six==1.11.0
stevedore==1.28.0
urllib3==1.22
warlock==1.3.0
wrapt==1.10.11
